refactor(nextjs): move provider/page-builder wiring to src/configuration/nextjs
Mirror ihexa's layout: framework wiring lives under src/configuration, not
src/libraries. Move the page builder providers configuration out of
src/libraries/nextjs/page into src/configuration/nextjs:

- client.ts  ClientBinder (the 'use client' provider component)
- index.ts   withClientBinder = createWithClientBinder(ClientBinder) and
             pageBuilder = createPageBuilder(withClientBinder)
- with-url-search-params.ts  the URL_SEARCH_PARAMS-providing middleware

Pages now import pageBuilder/withClientBinder/withUrlSearchParams from
@/configuration/nextjs. Delete src/libraries/nextjs/page entirely.

src/libraries/nextjs now only holds genuinely reusable, @arckit-agnostic
pieces: shim/ (Web Component router swap), href-with-search-params,
search.key and withStaticBearerAuth.

Allow src/configuration in .folderslintrc and as an app dependency target in
.dependency-cruiser.cjs.

(ClientBinder stays a local component rather than @arckit createClientBinder
because the latter is only exposed via @arckit/nextjs/client, which pulls
@arckit/daisyui — kept out to preserve the Web Component isolation.)

tsc clean, biome ci clean, architecture clean, 214 tests, Next + WC builds green.
