+++ src/app/(main)/(full-page)/(regions)/[region]/lieux/page.tsx
@@ -2,6 +2,7 @@ import { withFetch } from '@arckit/nextjs/page';
 import { withPagination, withSearchParams } from '@arckit/nextjs/page/middlewares';
 import type { Metadata } from 'next';
 import { notFound } from 'next/navigation';
+import { pageBuilder, withUrlSearchParams } from '@/configuration/nextjs';
 import { appendCollectivites } from '@/features/collectivites-territoriales';
 import { withRegion } from '@/features/collectivites-territoriales/middlewares/page';
 import { LieuxPage } from '@/features/lieux-inclusion-numerique';
@@ -10,7 +11,6 @@ import { type Region, regionMatchingSlug, regions } from '@/libraries/collectivi
 import { filtersSchema } from '@/libraries/inclusion-numerique-api';
 import { toLieuListItem } from '@/libraries/inclusion-numerique-api/transfer/to-lieu-list-item';
 import { hrefWithSearchParams } from '@/libraries/nextjs';
-import { pageBuilder, withUrlSearchParams } from '@/libraries/nextjs/page';
 import { appPageTitle, pageSchema } from '@/libraries/utils';
 
 type PageProps = {
