+++ src/app/(main)/(with-map)/(regions)/[region]/page.tsx
@@ -2,6 +2,7 @@ import { withFetch } from '@arckit/nextjs/page';
 import { withSearchParams } from '@arckit/nextjs/page/middlewares';
 import type { Metadata } from 'next';
 import { notFound } from 'next/navigation';
+import { pageBuilder } from '@/configuration/nextjs';
 import { DepartementsPage } from '@/features/cartographie';
 import { withRegion } from '@/features/collectivites-territoriales/middlewares/page';
 import { countLieux } from '@/features/lieux-inclusion-numerique/abilities/count/count-lieux';
@@ -13,7 +14,6 @@ import {
   regions
 } from '@/libraries/collectivites';
 import { filtersSchema } from '@/libraries/inclusion-numerique-api';
-import { pageBuilder } from '@/libraries/nextjs/page';
 import { contentId } from '@/libraries/ui/blocks/skip-links/skip-links';
 import SkipLinksPortal from '@/libraries/ui/blocks/skip-links/skip-links-portal';
 import { LocationFranceIllustration } from '@/libraries/ui/pictograms/map/location-france.illustration';
