+++ src/app/cartographie/regions/[region]/[departement]/[id]/page.tsx
@@ -1,5 +1,5 @@
+import { pageBuilder } from '@/configuration/nextjs';
 import { withLegacyDepartement, withLegacyId, withLegacyRegion } from '@/features/cartographie';
-import { pageBuilder } from '@/libraries/nextjs/page';
 
 export default pageBuilder()
   .use(withLegacyRegion())
