+++ src/app/cartographie/regions/[region]/page.tsx
@@ -1,5 +1,5 @@
+import { pageBuilder } from '@/configuration/nextjs';
 import { withLegacyRegion } from '@/features/cartographie';
-import { pageBuilder } from '@/libraries/nextjs/page';
 
 export default pageBuilder()
   .use(withLegacyRegion())
