+++ src/configuration/nextjs/client.ts
