+++ src/configuration/nextjs/index.ts
@@ -1,0 +1,8 @@
+import { createPageBuilder, createWithClientBinder } from '@arckit/nextjs/page';
+import { ClientBinder } from './client';
+
+export { withUrlSearchParams } from './with-url-search-params';
+
+export const withClientBinder = createWithClientBinder(ClientBinder);
+
+export const pageBuilder = createPageBuilder(withClientBinder);
