+++ src/configuration/nextjs/with-url-search-params.ts
