--- src/libraries/nextjs/page/middlewares/with-client-binder.tsx
@@ -1,4 +1,0 @@
-import { createWithClientBinder } from '@arckit/nextjs/page';
-import { ClientBinder } from './client-binder';
-
-export const withClientBinder = createWithClientBinder(ClientBinder);
